Factor out INode.Main
This moves are the basics common to MainNode that INode is interested in
into a a dedicated class. Also take FailedNode to the ride. This sets up
the state for #204.
